+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -4,10 +4,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.controller.CheckAdminInterceptor;
-import roomescape.controller.CheckLoginInterceptor;
 import roomescape.controller.LoginMemberArgumentResolver;
-import roomescape.controller.RequestTokenContext;
+import roomescape.controller.api.dto.request.TokenContextRequest;
 import roomescape.service.MemberService;
 import roomescape.util.TokenProvider;
 
@@ -17,26 +15,26 @@ import java.util.List;
 public class WebMvcConfiguration implements WebMvcConfigurer {
     private final MemberService memberService;
     private final TokenProvider tokenProvider;
-    private final RequestTokenContext requestTokenContext;
+    private final TokenContextRequest tokenContextRequest;
 
-    public WebMvcConfiguration(final MemberService memberService, final TokenProvider tokenProvider, final RequestTokenContext requestTokenContext) {
+    public WebMvcConfiguration(final MemberService memberService, final TokenProvider tokenProvider, final TokenContextRequest tokenContextRequest) {
         this.memberService = memberService;
         this.tokenProvider = tokenProvider;
-        this.requestTokenContext = requestTokenContext;
+        this.tokenContextRequest = tokenContextRequest;
     }
 
 
     @Override
     public void addInterceptors(final InterceptorRegistry registry) {
-        registry.addInterceptor(new CheckAdminInterceptor(memberService, tokenProvider, requestTokenContext))
+        registry.addInterceptor(new CheckAdminInterceptor(memberService, tokenProvider, tokenContextRequest))
                 .addPathPatterns("/admin/**");
-        registry.addInterceptor(new CheckLoginInterceptor(memberService,tokenProvider,requestTokenContext))
+        registry.addInterceptor(new CheckLoginInterceptor(memberService,tokenProvider, tokenContextRequest))
                 .addPathPatterns("/login/check")
                 .addPathPatterns("/reservations/**");
     }
 
     @Override
     public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new LoginMemberArgumentResolver(requestTokenContext));
+        resolvers.add(new LoginMemberArgumentResolver(tokenContextRequest));
     }
 }
